Trigger validation for publish and delete action, code style fix and remove useless variables

+++ src/Controller/JobController.php
@@ -83,7 +83,6 @@ class JobController extends AbstractController
         }
 
         return $this->render('job/create.html.twig', [
-            'job' => $job,
             'form' => $form->createView(),
         ]);
     }
@@ -159,7 +158,7 @@ class JobController extends AbstractController
         $form = $this->createDeleteForm($job);
         $form->handleRequest($request);
 
-        if ($form->isSubmitted()) {
+        if ($form->isSubmitted() && $form->isValid()) {
             $em->remove($job);
             $em->flush();
         }
@@ -184,7 +183,7 @@ class JobController extends AbstractController
         $form = $this->createPublishForm($job);
         $form->handleRequest($request);
 
-        if ($form->isSubmitted()) {
+        if ($form->isSubmitted() && $form->isValid()) {
             $job->setActivated(true);
 
             $em->flush();
